% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{analyticsutilities-package}
\alias{analyticsutilities}
\alias{analyticsutilities-package}
\title{analyticsutilities: Analytics Utility Functions}
\description{
A collection of utility functions for Philippine agricultural and
geospatial data analytics, including PSA PXWeb API wrappers, Supabase
REST helpers, PSGC geographic table builders, time-series extension,
and location-name standardisation.
}
\author{
\strong{Maintainer}: Nefriend Francisco \email{nefriendfrancisco@gmail.com}

}
\keyword{internal}
